Refactor Scene3D diagnostics helpers
