+++ workcell_builder/workcell_builder/gui/scene3d_diagnostics.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <QJsonArray>
+
+class QMatrix4x4;
+class QVector3D;
+
+// Scene3D diagnostics export helpers are implemented in scene3d_diagnostics.cpp.
+// The implementation is included from scene3d_viewport_widget.cpp as an incremental
+// refactor step while preserving existing private viewport state and behavior.
