+++ workcell_builder/workcell_builder/gui/scene3d_mesh_cache.h
@@ -1,0 +1,4 @@
+#pragma once
+
+// Incremental extraction target for Scene3D mesh loading, parsing, and cache helpers.
+// Behavior remains owned by scene3d_viewport_widget.cpp until the next small move.
