+++ workcell_builder/workcell_builder/gui/scene3d_transform_math.h
@@ -1,0 +1,4 @@
+#pragma once
+
+// Incremental extraction target for Scene3D transform and bounds math helpers.
+// Behavior remains owned by scene3d_viewport_widget.cpp until the next small move.
