+++ workcell_builder/workcell_builder/gui/scene3d_visual_classification.cpp
@@ -1,0 +1,1 @@
+#include "scene3d_visual_classification.h"
