+++ workcell_builder/workcell_builder/gui/scene3d_visual_classification.h
@@ -1,0 +1,4 @@
+#pragma once
+
+// Incremental extraction target for Scene3D visual classification/filtering helpers.
+// Behavior remains owned by scene3d_viewport_widget.cpp until the next small move.
